requirements.txt
# Python dependencies
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=10.0.0
tqdm>=4.66.0
python-multipart>=0.0.6

# For theoretical notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
